backend/requirements.txt
# Python >= 3.11
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
sqlalchemy==2.0.27
pydantic==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
alembic==1.13.1
psycopg2-binary==2.9.9 